Fix issue #159: Support context configured with cookies="false"
A new session was only stored in memcached if there was a session cookie sent
to the client. Thus cookies="false" was not supported.

This is fixed by storing a note at the request that provides the session id
if a session was created.
When the request came in with a requestedSessionId and it will be / was changed
(in the case of a tomcat/memcached failover) we don't need to note the new/changed
session id because the request.changeSessionId also changes the requestedSessionId
which is also used for getting the sessionId for the session backup.