Resolve #168: Add tomcat8 support.
Thanx @h0nIg (Hans-Joachim Kliemeck) and @cyrille-leclerc (Cyrille Le Clerc).

See also pull requests #28 and #30

Tomcat8 now redirects with 303 instead of 302. For 303 httpclient automatically
redirects, so httpclient must be asked not to do so, additionally the status code
checks in TestUtils needed to be extended for 303.

Change context path to "" to get rid of tomcat8 //login workaround