Fix #190 updateExpirationInMemcached prints negative memcachedExpirationTime
This was caused by tomcats internal "processExpiresFrequency": tomcat doesn't expire()
sessions as soon as they reached their timeout, but they're expired according to the
configured processExpiresFrequency, which is 6 by default. Because tomcats background
process runs every 10 seconds, and processExpiresFrequency is 6 by default, a session
with timeout 5 minutes will be exipire'd 60 seconds after the session reached its timeout.
This lead to the negative memcachedExpirationTime.

With this fix the expiration time that is set for sessions in memcached respects the
offset caused by processExpiresFrequency so that msm should no longer print any negative
memcachedExpirationTime.