Added option to use Kryo's CompatibleFieldSerializer
This patch introduces the new boolean parameter
msm.kryo.compatible. If this property is set to true Kryo will
use CompatibleFieldSerializer as the default serializer instead
of the ReferenceFieldSerializer which is configured in
KryoReflectionFactorySupport.

This can be used to create some compatibility between slightly
modified classes in memcached, e. g. when having minor updates to
your software.