Fix invocation of StandardSession#exclude method
Fixed invocation of StandardSession#exclude method so that it works even
with older Tomcat versions where only the one parameter version of this
method was present.

If only the legacy one parameter method is detected, it is invoked via
reflection, otherwise the new two parameter variant is invoked in
standard manner.

This fix was tested with Tomcat 8.0.14 (which only has the one parameter
exclude method) and 8.0.36 (which has both one and two parameter exclude
methods).

Fixes #303