Gracefully handle exceptions during deletion of backup data
This was reported in #322. `LockingStrategy.onAfterDeleteFromMemcached`
could fail if a backup node was unavailable, which caused `SuffixBasedNodeLocator.getPrimary`
to throw an IAE. The latter is now changed to not only take _available_ nodes
into account but all primary nodes (including unavailable nodes). This
should reduce the occurrences of the situation already.

Additionally `LockingStrategy.onAfterDeleteFromMemcached` now catches
exceptions for deletion of backup data and logs an info, as this is considered
to be a non-critical case (memcached will evict data eventually).

Refs #322